Adjustements for building with VS2017.
