+++ src/openboardview/utils.cpp
@@ -10,7 +10,6 @@
 #include <sstream>
 #include <stdint.h>
 
-#include <dirent.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 
@@ -57,22 +56,6 @@ bool compare_string_insensitive(const std::string &str1, const std::string &str2
 	return str1.size() == str2.size() && std::equal(str2.begin(), str2.end(), str1.begin(), [](const char &a, const char &b) {
 		       return std::tolower(a) == std::tolower(b);
 		   });
-}
-
-// Case insensitive lookup of a filename at the given path
-std::string lookup_file_insensitive(const std::string &path, const std::string &filename) {
-	std::string filefound;
-	DIR *dir;
-	struct dirent *dent;
-
-	dir = opendir(path.c_str()); /* any suitable directory name  */
-	if (!dir) return filefound;
-
-	while ((dent = readdir(dir)) != NULL) {
-		std::string cfile(dent->d_name);
-		if (compare_string_insensitive(cfile, filename)) filefound = path + cfile;
-	}
-	return filefound;
 }
 
 // Split a string in a vector, delimiter is a space (stringstream iterator)
