+++ src/openboardview/utils.h
@@ -16,8 +16,5 @@ bool find_str_in_buf(const std::string str, const std::vector<char> &buf);
 // Case insensitive comparison of std::string
 bool compare_string_insensitive(const std::string &str1, const std::string &str2);
 
-// Case insensitive lookup of a filename at the given path
-std::string lookup_file_insensitive(const std::string &path, const std::string &filename);
-
 // Split a string in a vector, delimiter is a space (stringstream iterator)
 std::vector<std::string> split_string(const std::string str);
